requirements.txt
apted==1.0.3
beautifulsoup4==4.12.2
Distance==0.1.3
lxml==4.9.2
matplotlib==3.7.1
mmcv_full==1.7.0
mmdet==2.28.2
numpy==1.24.3
opencv_python==4.7.0.72
pandas==2.0.2
Pillow==8.2.0
Pillow==10.0.0
Polygon==1.1.0
Polygon3==3.0.9.1
seaborn==0.12.2
Shapely==2.0.1
torch==1.12.0
torchvision==0.13.0
tqdm==4.65.0